Added permanently deleted function
Internally we change the deleted_at date of a permanently deleted message to more than MESSAGES_DELETED_MAX_AGE days ago.
If both users permanently deleted a message or if deleted_at is more than MESSAGES_DELETED_MAX_AGE days ago, the message is deleted from the database.
Even if the message could not be deleted from the database (one of the users did not delete the message), the message will not appear in the trash anymore.